requirements.txt
python-telegram-bot==20.0
gspread==6.1.4
oauth2client==4.1.3
python-dotenv==1.0.0
requests==2.31.0
gspread-formatting==1.2.0